Add a "public interface" for the certs generator
The UI is just a binary called with a long series of command-line args.
The list of participants to generate certificates for is a simple TSV
one. Emails are always downcased for consistency. Names are
transliterated to ASCII to avoid unicode issues with LaTeX and to make
the certificates readable in English.

The certificate file name is moved to the bin file as it's not a concern
of the PDF generator.